Check for permissions every 5s during the game
